cld80211lib: Remove sendmsg to driver during cleanup
The get_multicast_id initiates a sendmsg to the driver and waits
in busywait to receive the msg. In certain failures where driver
is stopped prior to cleanup, the error handler is not returning
to terminate the bsywait. Subsequently we don't send a cleanup
notification to the framework. This triggers a subsequent start
of HAL interface and a hang state.

Bug: 64350646
Change-Id: I4cb078a98830f667beceb78a356cd7a4808b2f73
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>